requirements.txt
langchain==0.3.14
langchain-openai==0.2.14
langchain-community==0.3.14
langchain-core==0.3.22
python-dotenv==1.0.0